LayoutTests/platform/qt-wk1/fast/loader/willsendrequest-returns-null-for-memory-cache-load-expected.txt

resources/compass.jpg - willSendRequest <NSURLRequest URL resources/compass.jpg, main document URL willsendrequest-returns-null-for-memory-cache-load.html, http method GET> redirectResponse (null)
<unknown> - didFinishLoading
resources/compass.jpg - didReceiveResponse <NSURLResponse resources/compass.jpg, http status code 0>
resources/compass.jpg - didFinishLoading
resources/cached-image.html - willSendRequest <NSURLRequest URL resources/cached-image.html, main document URL willsendrequest-returns-null-for-memory-cache-load.html, http method GET> redirectResponse (null)
resources/cached-image.html - didReceiveResponse <NSURLResponse resources/cached-image.html, http status code 0>
resources/compass.jpg - willSendRequest <NSURLRequest URL resources/compass.jpg, main document URL , http method GET> redirectResponse (null)
resources/compass.jpg - didReceiveResponse <NSURLResponse resources/compass.jpg, http status code 0>
resources/compass.jpg - didFinishLoading
CONSOLE MESSAGE: line 5: FAIL: image was incorrectly loaded
  
